backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tensorflow==2.20.0
pillow==10.2.0
numpy==1.24.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python==4.8.1.78
aiofiles==23.2.1
pymongo[srv]==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==42.0.0
requests==2.31.0
dnspython==2.5.0


numpy==1.26.3